requirements.txt
fastapi
uvicorn
tensorflow
numpy
Pillow
matplotlib
fpdf
qrcode
deep-translator
pymongo
